requirements.txt
Flask==2.1.2
Werkzeug==2.2.2
google-cloud-aiplatform==1.25.0